app/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
exabgp==4.2.17
idna==3.3
Jinja2==3.1.1
MarkupSafe==2.1.1
PyYAML==6.0
requests==2.27.1
urllib3==1.26.9
